+++ chat.py
@@ -1,0 +1,6 @@
+#chat
+
+class Chat:
+
+    def response(str):
+        return str
